requirements.txt
streamlit==1.31.1
openai==1.12.0
python-dotenv==1.0.1
oss2==2.18.3
pandas==2.2.0
matplotlib==3.8.2
networkx==3.2.1
pyvis==0.3.2
jinja2==3.1.3
pdfplumber==0.10.3
python-docx==1.1.0
Pillow==10.2.0
torch==2.2.0
numpy==1.26.4
requests==2.31.0
scikit-learn==1.4.0
tqdm==4.66.2
